4.21 done

+++ osa4/blogilista/controllers/blogs.js
@@ -44,7 +44,18 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() !== decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    } else {
+        await Blog.findByIdAndDelete(request.params.id)
+    }
     response.status(204).end()
 })
